Display audio of WebM files in waveform
WebM is just a variant of Matroska, but it has its own mime-type.
Hence it must be enabled separately. I added mime-types for tagged
MP3 files, too.

I guess this commit shows why the selection of the correct stream
should be improved. However, most files should work now so this
so this can be considered a temporary fix.